Migrate search targets to new pattern
